bound Flood_Fill to Python
